+++ messages/output.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"math"
 	"os"
+	"regexp"
 	"strings"
 
 	"github.com/UpCloudLtd/progress/terminal"
@@ -17,6 +18,8 @@ type RenderState int
 const (
 	RenderStateDone RenderState = -1
 )
+
+var whitespace = regexp.MustCompile(`\s`)
 
 type OutputConfig struct {
 	DefaultTextWidth            int
@@ -171,7 +174,14 @@ func (cfg OutputConfig) GetMaxHeight() int {
 
 func (cfg OutputConfig) formatDetails(msg *Message) string {
 	wrapWidth := cfg.GetMaxWidth() - 2
-	details := text.WrapSoft(cfg.getDetailsColor().Sprint(msg.Details), wrapWidth)
+
+	var details string
+	// If details contains newline characters, assume that details are preformatted (e.g., stack trace, console output, ...)
+	if strings.Contains(msg.Details, "\n") {
+		details = text.WrapText(cfg.getDetailsColor().Sprint(msg.Details), wrapWidth)
+	} else {
+		details = text.WrapSoft(cfg.getDetailsColor().Sprint(msg.Details), wrapWidth)
+	}
 
 	if cfg.ShowStatusIndicator {
 		return strings.ReplaceAll("\n"+details, "\n", "\n  ")
@@ -206,6 +216,7 @@ func (cfg OutputConfig) GetMessageText(msg *Message, renderState RenderState) st
 	if maxMessageWidth < 0 {
 		return ""
 	}
+	message = whitespace.ReplaceAllString(message, " ")
 	if len(message) > maxMessageWidth {
 		message = fmt.Sprintf("%s…", message[:maxMessageWidth-1])
 	} else {